# api/docs/swagger.yaml
basePath: /
definitions:
  api.apiError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
      type:
        type: string
    type: object
  api.apiErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.apiError'
      object:
        description: '"error"'
        type: string
    type: object
...
  api.diffResponse:
    properties:
      created:
        type: integer
      document_id:
        type: string
      id:
        type: string
      left_file_id:
        type: string
      object:
        type: string
      right_file_id:
        type: string
      similarity_percent:
        type: number
      status:
        type: string
    type: object
...
  api.listResponse-api_diffResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.diffResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
      object:
        description: '"list"'
        type: string
    type: object
...
host: legist.nadevko.cc
info:
  contact: {}
  description: |-
    AI assistant for comparing revisions of normative legal acts (НПА) of the Republic of Belarus.

    ## Authentication
    All endpoints except public auth routes require a Bearer token in the `Authorization` header.
    Obtain tokens via `POST /sessions`. Refresh via `POST /tokens/refresh`.

    ## Request IDs
    Every response includes a `Request-Id` header for debugging.

    ## Idempotency
    `POST` requests require an `Idempotency-Key` header.
    Repeating a request with the same key returns the cached response.
    Keys expire after 24 hours. Reusing a key for a different endpoint returns 422.

    ## Roles
    Users have `role`: `user` or `admin` (in `users` table and JWT `role` claim).
    Only `admin` may change another user's role via `PATCH /users/:id`.
    In dev (`ENV=dev`), new registrations default to `admin`; otherwise default is `user`.

    ## Ownership and access control
    Resources belong to the authenticated user. Attempting to read or mutate
    another user's private resource returns 404 (not 403) to avoid leaking existence.
    Public resources (`user_id IS NULL`) are readable by any authenticated user but cannot be mutated.
    List `owner` query (Stripe-like): omit → non-admin sees only own rows; admin sees own + public.
    `owner=null` or `owner=public` (aliases, admin only) → public rows only. `owner=<your user id>` → own rows only. Other ids are rejected.

...

...

...

...

...

...

...

...

...

...
paths:
...
  /diffs:
    get:
      parameters:
      - description: Filter by document
        in: query
        name: document_id
        type: string
      - description: Filter by left or right file
        in: query
        name: file_id
        type: string
      - description: pending|processing|done|failed
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: starting_after
        type: string
      - description: Cursor
        in: query
        name: ending_before
        type: string
      - collectionFormat: csv
        description: 'Expand: document, left_file, right_file'
        in: query
        items:
          type: string
        name: expand[]
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listResponse-api_diffResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: List diffs for the current user
      tags:
      - Diffs
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 'With right_file_id: compare two existing files'
        in: formData
        name: left_file_id
        type: string
      - description: 'With left_file_id: compare two existing files'
        in: formData
        name: right_file_id
        type: string
      - description: 'With left_file_id or right_file_id: new version to compare'
        in: formData
        name: file
        type: file
      - description: 'With file_right: create document and compare'
        in: formData
        name: file_left
        type: file
      - description: 'With file_left: create document and compare'
        in: formData
        name: file_right
        type: file
      - description: Work metadata when creating document (two files)
        in: formData
        name: subtype
        type: string
      - description: Work metadata
        in: formData
        name: number
        type: string
      - description: Work metadata
        in: formData
        name: author
        type: string
      - description: Work metadata
        in: formData
        name: date
        type: string
      - description: Work metadata
        in: formData
        name: country
        type: string
      - description: Work metadata
        in: formData
        name: name
        type: string
      - description: application/json | text/event-stream
        in: header
        name: Accept
        type: string
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.diffResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: 'Create a diff (multipart: two file IDs, ID + file, or two files)'
      tags:
      - Diffs
  /diffs/{id}:
    get:
      parameters:
      - description: Diff ID
        in: path
        name: id
        required: true
        type: string
      - collectionFormat: csv
        description: 'Expand: document, left_file, right_file'
        in: query
        items:
          type: string
        name: expand[]
        type: array
      - description: application/json | text/event-stream
        in: header
        name: Accept
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.diffResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get diff by ID or stream pipeline (SSE)
      tags:
      - Diffs
...
securityDefinitions:
  BearerAuth:
    description: 'Enter token in format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...